# Allow scientific notation in written numbers