--- a/src/write.jl
+++ b/src/write.jl
@@ -2,107 +2,41 @@
 ##### utilities
 #####
 
-# `allow_scientific` is only meaningful for `value::Number`. We allow passing it though,
-# so `edf_write` can be more generic.
-_edf_repr(value::Union{String,Char}; allow_scientific=nothing) = value
-_edf_repr(date::Date; allow_scientific=nothing) = uppercase(Dates.format(date, dateformat"dd-u-yyyy"))
-_edf_repr(date::DateTime; allow_scientific=nothing) = Dates.format(date, dateformat"dd\.mm\.yyHH\.MM\.SS")
+const FORMATS = (; G=[Printf.Format("%.$(i)G") for i in 8:-1:1],
+                 F=[Printf.Format("%.$(i)F") for i in 8:-1:1])
 
-"""
-    sprintf_G_under_8(x) -> String
-
-Return a string of length at most 8, written in either scientific notation (using capital-'E'),
-or in decimal format, using as much precision as possible.
-"""
-function sprintf_G_under_8(x)
-    shorten = str -> begin
-        if contains(str, 'E')
-            # Remove extraneous 0's in exponential notation
-            str = replace(str, "E-0" => "E-")
-            str = replace(str, "E+0" => "E+")
-            # Remove any leading/trailing whitespace
-            return strip(str)
-        else
-            # Decimal format. Call out to `trim_F`
-            return trim_F(str)
+function _edf_repr(x::Real, allow_scientific::Bool=false)
+    fmts = allow_scientific ? FORMATS.G : FORMATS.F
+    for fmt in fmts
+        str = Printf.format(fmt, x)
+        # Remove extra zero in scientific notation, e.g. `1E+01` becomes `1E+1`
+        # Also drop decimal if all digits are zero: `123.0000` to `123`
+        str = replace(str, r"(E[+-])0" => s"\1", r"\.0+$" => "")
+        # Removing trailing 0's after decimal place
+        if contains(str, '.') && !allow_scientific
+            str = rstrip(str, '0')
+        end
+        str = strip(str)
+        if length(str) <= 8
+            if str == "0" && x != 0
+                @warn "Underflow to zero when writing number `$x` to 8-character ASCII" maxlog=10
+            end
+            return String(str)
         end
     end
-    # Strategy:
-    # `@printf("%0.NG", x)` means:
-    # - `G`: Use the shortest representation: %E or %F. That is, scientific notation (with capital E) or decimals, whichever is shorted
-    # - `.N`: (for literal `N`, like `5`): the maximum number of significant digits to be printed
-    # However, `@printf("%0.NG", x)` may have more than `N` characters (e.g. presence of E, and the values for the exponent, the decimal place, etc)
-    # So we start from 8 (most precision), and stop as soon as we get under 8 characters
-    sig_8 = shorten(@sprintf("%.8G", x))
-    length(sig_8) <= 8 && return sig_8
-    sig_7 = shorten(@sprintf("%.7G", x))
-    length(sig_7) <= 8 && return sig_7
-    sig_6 = shorten(@sprintf("%.6G", x))
-    length(sig_6) <= 8 && return sig_6
-    sig_5 = shorten(@sprintf("%.5G", x))
-    length(sig_5) <= 8 && return sig_5
-    sig_4 = shorten(@sprintf("%.4G", x))
-    length(sig_4) <= 8 && return sig_4
-    sig_3 = shorten(@sprintf("%.3G", x))
-    length(sig_3) <= 8 && return sig_3
-    sig_2 = shorten(@sprintf("%.2G", x))
-    length(sig_2) <= 8 && return sig_2
-    sig_1 = shorten(@sprintf("%.1G", x))
-    length(sig_1) <= 8 && return sig_1
-    error("failed to fit number into EDF's 8 ASCII character limit: $x")
+    throw(ArgumentError("cannot represent $x in 8 ASCII characters"))
 end
 
-function trim_F(str)
-    if contains(str, '.')
-        # Remove trailing 0's after the decimal point
-        str = rstrip(str, '0')
-        # If the `.` is at the end now, strip it
-        str = rstrip(str, '.')
-    end
-    # Removing leading or trailing whitespace
-    str = strip(str)
-    return str
-end
+_edf_repr(value, ::Any) = _edf_repr(value)
 
-function sprintf_F_under_8(x)
-    shorten = trim_F
-    # Strategy:
-    # `@printf("%0.NF", x)` means:
-    # - `F`: print with decimals
-    # - `.N`: (for literal `N`, like `5`): the maximum number of digits to print after the decimal
-    # However, `@printf("%0.NF", x)` may have more than `N` characters (e.g. digits to the left of the decimal point)
-    # So we start from 8 (most precision), and stop as soon as we get under 8 characters
-    sig_8 = shorten(@sprintf("%.8F", x))
-    length(sig_8) <= 8 && return sig_8
-    sig_7 = shorten(@sprintf("%.7F", x))
-    length(sig_7) <= 8 && return sig_7
-    sig_6 = shorten(@sprintf("%.6F", x))
-    length(sig_6) <= 8 && return sig_6
-    sig_5 = shorten(@sprintf("%.5F", x))
-    length(sig_5) <= 8 && return sig_5
-    sig_4 = shorten(@sprintf("%.4F", x))
-    length(sig_4) <= 8 && return sig_4
-    sig_3 = shorten(@sprintf("%.3F", x))
-    length(sig_3) <= 8 && return sig_3
-    sig_2 = shorten(@sprintf("%.2F", x))
-    length(sig_2) <= 8 && return sig_2
-    sig_1 = shorten(@sprintf("%.1F", x))
-    length(sig_1) <= 8 && return sig_1
-    error("failed to fit number into EDF's 8 ASCII character limit: $x")
-end
-
-function _edf_repr(x::Real; allow_scientific=false)
-    if allow_scientific
-        return sprintf_G_under_8(x)
-    else
-        return sprintf_F_under_8(x)
-    end
-end
+_edf_repr(value::Union{String,Char}) = value
+_edf_repr(date::Date) = uppercase(Dates.format(date, dateformat"dd-u-yyyy"))
+_edf_repr(date::DateTime) = Dates.format(date, dateformat"dd\.mm\.yyHH\.MM\.SS")
 
 _edf_metadata_repr(::Missing) = 'X'
 _edf_metadata_repr(x) = _edf_repr(x)
 
-function _edf_repr(metadata::T; allow_scientific=nothing) where {T<:Union{PatientID,RecordingID}}
+function _edf_repr(metadata::T) where {T<:Union{PatientID,RecordingID}}
     header = T <: RecordingID ? String["Startdate"] : String[]
     # None of the fields of `PatientID` or `RecordingID` are floating point, so we don't need
     # to worry about passing `allow_scientific=true`.
@@ -110,7 +44,7 @@ function _edf_repr(metadata::T; allow_scientific=nothing) where {T<:Union{Patien
 end
 
 function edf_write(io::IO, value, byte_limit::Integer; allow_scientific=false)
-    edf_value = _edf_repr(value; allow_scientific)
+    edf_value = _edf_repr(value, allow_scientific)
     sizeof(edf_value) > byte_limit && error("EDF value exceeded byte limit (of $byte_limit bytes) while writing: `$value`. Representation: `$edf_value`")
     bytes_written = Base.write(io, edf_value)
     while bytes_written < byte_limit