--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,6 +1,6 @@
 using EDF
 using EDF: TimestampedAnnotationList, PatientID, RecordingID, SignalHeader,
-           Signal, AnnotationsSignal
+           Signal, AnnotationsSignal, sprintf_G_under_8, sprintf_F_under_8
 using Dates
 using FilePathsBase
 using Test
@@ -40,6 +40,43 @@ deep_equal(a::T, b::S) where {T,S} = false
 const DATADIR = joinpath(@__DIR__, "data")
 
 @testset "Just Do It" begin
+
+    @testset "sprintf functionality" begin
+        # Moderate sized numbers
+        @test sprintf_G_under_8(0.123) == "0.123"
+        @test sprintf_F_under_8(0.123) == "0.123"
+
+        @test sprintf_G_under_8(1.123) == "1.123"
+        @test sprintf_F_under_8(1.123) == "1.123"
+
+        @test sprintf_G_under_8(10.123) == "10.123"
+        @test sprintf_F_under_8(10.123) == "10.123"
+
+        # Moderate numbers, many digits
+        @test sprintf_G_under_8(0.8945620050698592) == "0.894562"
+        @test sprintf_F_under_8(0.8945620050698592) == "0.894562"
+
+        # Large numbers / few digits
+        @test sprintf_G_under_8(0.123e10) == "1.23E+9"
+        # decimal version cannot handle it:
+        err = ErrorException("failed to fit number into EDF's 8 ASCII character limit: 1.23e9")
+        @test_throws err sprintf_F_under_8(0.123e10)
+
+        # Large numbers / many digits
+        @test sprintf_G_under_8(0.8945620050698592e10) == "8.946E+9"
+        err = ErrorException("failed to fit number into EDF's 8 ASCII character limit: 8.945620050698591e9")
+        @test_throws err sprintf_F_under_8(0.8945620050698592e10)
+
+        # Small numbers / few digits
+        @test sprintf_G_under_8(0.123e-10) == "1.23E-11"
+        # decimal version underflows:
+        @test sprintf_F_under_8(0.123e-10) == "0"
+
+        # Small numbers / many digits
+        @test sprintf_G_under_8(0.8945620050698592e-10) == "8.95E-11"
+        @test sprintf_F_under_8(0.8945620050698592e-10) == "0"
+    end
+
     # test EDF.read(::AbstractString)
     edf = EDF.read(joinpath(DATADIR, "test.edf"))
     @test sprint(show, edf) == "EDF.File with 140 16-bit-encoded signals"
@@ -68,7 +105,7 @@ const DATADIR = joinpath(@__DIR__, "data")
                         [TimestampedAnnotationList(4.0, nothing, String[""]), TimestampedAnnotationList(2.5, 2.5, ["type A"])],
                         [TimestampedAnnotationList(5.0, nothing, String[""])]]
             @test all(signal.records .== expected)
-            @test AnnotationsSignal(signal.records).samples_per_record == 16
+            @test AnnotationsSignal(signal.records).samples_per_record == 14
         end
     end
 
@@ -113,19 +150,22 @@ const DATADIR = joinpath(@__DIR__, "data")
     end
 
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(-0.0023405432)) == "-0.00234"
-    @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(0.0023405432)) == "0.002340"
+    @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(0.0023405432)) == "0.002341"
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(1.002343)) == "1.002343"
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(1011.05432)) == "1011.054"
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(-1011.05432)) == "-1011.05"
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(-1013441.5)) == "-1013442"
     @test EDF._edf_repr(EDF._nearest_representable_edf_time_value(-1013441.3)) == "-1013441"
     @test EDF._edf_repr(34577777) == "34577777"
-    @test EDF._edf_repr(0.0345) == "0.034500"
-    @test EDF._edf_repr(-0.02) == "-0.02000"
+    @test EDF._edf_repr(0.0345) == "0.0345"
+    @test EDF._edf_repr(-0.02) == "-0.02"
     @test EDF._edf_repr(-187.74445) == "-187.744"
     @test_throws ErrorException EDF._edf_repr(123456789)
     @test_throws ErrorException EDF._edf_repr(-12345678)
-    @test_throws ErrorException EDF._edf_repr(0.00000000024)
+
+    # if we don't allow scientific notation, we allow rounding down here
+    @test EDF._edf_repr(0.00000000024) == "0"
+    @test EDF._edf_repr(0.00000000024; allow_scientific=true) == "2.4E-10"
     @test_throws ErrorException EDF.edf_write(IOBuffer(), "hahahahaha", 4)
 
     uneven = EDF.read(joinpath(DATADIR, "test_uneven_samp.edf"))
@@ -246,10 +286,10 @@ end
 
 @testset "BDF+ Files" begin
     # This is a `BDF+` file containing only trigger information.
-    # It is similiar to a `EDF Annotations` file except that 
+    # It is similiar to a `EDF Annotations` file except that
     # The `ANNOTATIONS_SIGNAL_LABEL` is `BDF Annotations`.
-    # The test data has 1081 trigger events, and 
-    # has 180 trials in total, and 
+    # The test data has 1081 trigger events, and
+    # has 180 trials in total, and
     # The annotation `255` signifies the offset of a trial.
     # More information, contact: zhanlikan@hotmail.com
     evt = EDF.read(joinpath(DATADIR, "evt.bdf"))