Update to #30: attempts to allow writing out scientific notation where legal (i.e. not onset or duration fields).

motivated by encountering
```julia
ErrorException(\"failed to fit number into EDF's 8 ASCII character limit: 4.180821e-7\")
```
when writing out an EDF. We don't know why or which field this was in though.

This PR changes the semantics:
* if `allow_scientific=false`, which I have set by default, then *underflow to 0 is allowed*. This was formerly disallowed in #30 (although I did not see a discussion why). The former semantics were a bit weird in that rounding/truncation was OK except when the result was truncated to 0, in which case it errored.
* if `allow_scientific=true`, then we use `%G` in sprintf to allow either of scientific notation or decimal representation. This shouldn't really underflow nor with any kind of Float64 representable number, but could if somehow one got a BigFloat or such in there.

I also set the code to allow scientific notation with all floating-point fields of the headers (but with onset/durations of annotations).

Would appreciate thoughts on this & careful review