@ericphanson 2023-10-13
Is this too slow?

* the `allow_scientific=true` path is only used for headers, so should not be a perf concern
* the `allow_scientific=false` path is used for annotations, so could be a perf concern, if it is slow to call for several million numbers

Let us benchmark.
## Benchmark
Try every Float16 50 times, which is 1.5 million numbers
```julia
function all_floats(; n=50)
    out = Char[]
    sizehint!(out, 30720 * n)
    for _ = 1:n
        x = floatmin(Float16)
        while x <= floatmax(Float16)
            str = EDF._edf_repr(Float64(x))
            push!(out, str[1])
            x = nextfloat(x)
        end
    end
    return out
end
```
## PR
```julia
julia> @time length(all_floats())
  0.731282 seconds (9.98 M allocations: 1.442 GiB, 7.50% gc time)
1536000

julia> @time length(all_floats())
  0.727198 seconds (9.98 M allocations: 1.442 GiB, 7.72% gc time)
1536000

julia> @time length(all_floats())
  0.736517 seconds (9.98 M allocations: 1.442 GiB, 7.93% gc time)
1536000
```

## Master
```julia
julia> @time length(all_floats())
  0.344242 seconds (12.52 M allocations: 889.876 MiB, 6.08% gc time)
1536000

julia> @time length(all_floats())
  0.345693 seconds (12.52 M allocations: 889.876 MiB, 5.95% gc time)
1536000

julia> @time length(all_floats())
  0.347998 seconds (12.52 M allocations: 889.876 MiB, 6.80% gc time)
1536000
```

## Conclusion
So it is ~2x slower, but still fast enough that it doesn't seem like a big issue.