@ericphanson 2023-10-13
## Is allowing underflow to 0 dangerous?

* We allow this now when `allow_scientific=false`. This is used for onsets & annotations, and for non-floating-point types.
    * for onsets and annotations, a sufficeintly small time like `1e-7` does seem OK to round down to 0. If the onset is right at the start of the recording, marking that as time 0 seems OK. If an annotation has very small duration, since EDF does allow 0 or even missing duration, rounding that down to 0 does seem OK too.
* when `allow_scientific=true`, we can represent even `floatmin(Float64)` with a few digits of precision. This is used for floating point fields in the headers.

Therefore I think it is probably OK, given the we can see all the usages and check them individually.