@ericphanson 2023-10-13
## Is scientific notation even allowed in the EDF+ spec?

from https://www.edfplus.info/specs/guidelines.html#:~:text=Numbers%20in%20EDF(%2B)%20headers,3%20(Okt%202004) they say

> 2a, E-notation (May 2009). Numbers in EDF(+) headers may have the scientific E notation as in 1E2345, +012E+34, -1.34E09 and +1.234E-5. Note, though, that the 8 characters of some EDF(+) numberfields are used more efficiently by -123.456 uV than by -1.23E-4 V.

So it seems it is. We do not use it for non-header fields such as annotation durations and onsets.