@ericphanson 2023-10-16
One thing I realized is that in TAL, there isn't the same 8 character limits (as far as I can tell in the spec). And that is the only place we currently force the decimal printing with floating-point numbers. So we could try raising the limit there (or just do a regular `@sprintf("%f", x)`).