@ericphanson 2023-10-17
This actually does fix my internal bug, which now has an MWE at https://github.com/beacon-biosignals/EDF.jl/issues/74, and results in EDFs that are loadable in EDFBrowser at least (and which pass it's validation check) but I don't think this approach is actually necessary because we should just allow rounding to 0 for fields like `physical_min`. So I think a different fix would be more direct here as well as less concerning about portability.